requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
dash
plotly